+++ packages/kiapi/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.5.1] - 2026-07-29
+
 ### Added
 
 - `kiapi run --relay none` explicitly disables the relay, overriding a relay enabled in user settings.